requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
azure-identity
azure-keyvault-secrets
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
distlib==0.3.8
distro==1.9.0
Django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.4.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
multidict==6.0.5
numpy==1.26.3
openai==0.28.0
pandas==2.2.0
pipenv==2023.11.15
platformdirs==4.1.0
psycopg2
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
virtualenv==20.25.0
whitenoise==6.4.0
yarl==1.9.4
